✅ FEAT: make ActionofPartner page a component
